refactor(plugin-rem): use modifyHTMLTags hook
