+++ packages/plugin-rem/src/index.ts
@@ -1,26 +1,52 @@
-import type { PostCSSPlugin, RsbuildPlugin } from '@rsbuild/core';
+import path from 'node:path';
+import {
+  type PostCSSPlugin,
+  type RsbuildPlugin,
+  ensureAssetPrefix,
+  logger,
+} from '@rsbuild/core';
 import deepmerge from 'deepmerge';
+import { DEFAULT_OPTIONS, getRootPixelCode } from './helpers';
 import type { PluginRemOptions, PxToRemOptions } from './types';
-
-const defaultOptions: PluginRemOptions = {
-  enableRuntime: true,
-  rootFontSize: 50,
-};
 
 export type { PluginRemOptions };
 
 export const PLUGIN_REM_NAME = 'rsbuild:rem';
 
-export const pluginRem = (options: PluginRemOptions = {}): RsbuildPlugin => ({
+export const pluginRem = (
+  userOptions: PluginRemOptions = {},
+): RsbuildPlugin => ({
   name: PLUGIN_REM_NAME,
 
   setup(api) {
-    const userOptions = {
-      ...defaultOptions,
-      ...options,
+    const options = {
+      ...DEFAULT_OPTIONS,
+      ...userOptions,
     };
 
-    const getPxToRemPlugin = async () => {
+    let scriptPath: string | undefined;
+    let runtimeCode: string | undefined;
+
+    const getScriptPath = (distDir: string) => {
+      if (!scriptPath) {
+        scriptPath = path.posix.join(
+          distDir,
+          `convert-rem.${RSBUILD_VERSION}.js`,
+        );
+      }
+
+      return scriptPath;
+    };
+
+    const getRuntimeCode = async () => {
+      if (!runtimeCode) {
+        const isCompress = process.env.NODE_ENV === 'production';
+        runtimeCode = await getRootPixelCode(options, isCompress);
+      }
+      return runtimeCode;
+    };
+
+    const getPostCSSPlugin = async () => {
       const { default: pxToRemPlugin } = (await import(
         '../compiled/postcss-pxtorem/index.js'
       )) as {
@@ -28,15 +54,15 @@ export const pluginRem = (options: PluginRemOptions = {}): RsbuildPlugin => ({
       };
 
       return pxToRemPlugin({
-        rootValue: userOptions.rootFontSize,
+        rootValue: options.rootFontSize,
         unitPrecision: 5,
         propList: ['*'],
-        ...(userOptions.pxtorem ? deepmerge({}, userOptions.pxtorem) : {}),
+        ...(options.pxtorem ? deepmerge({}, options.pxtorem) : {}),
       });
     };
 
     api.modifyRsbuildConfig(async (config, { mergeRsbuildConfig }) => {
-      const remPlugin = await getPxToRemPlugin();
+      const remPlugin = await getPostCSSPlugin();
       return mergeRsbuildConfig(config, {
         tools: {
           postcss(_, { addPlugins }) {
@@ -46,29 +72,90 @@ export const pluginRem = (options: PluginRemOptions = {}): RsbuildPlugin => ({
       });
     });
 
-    api.modifyBundlerChain(
-      async (chain, { target, CHAIN_ID, environment, HtmlPlugin }) => {
-        if (target !== 'web' || !userOptions.enableRuntime) {
-          return;
+    api.modifyBundlerChain(async (chain, { target, CHAIN_ID, environment }) => {
+      if (target !== 'web' || !options.enableRuntime || options.inlineRuntime) {
+        return;
+      }
+
+      const { AutoSetRootFontSizePlugin } = await import('./helpers');
+      const { config } = environment;
+
+      chain
+        .plugin(CHAIN_ID.PLUGIN.AUTO_SET_ROOT_SIZE)
+        .use(AutoSetRootFontSizePlugin, [
+          getScriptPath(config.output.distPath.js),
+          getRuntimeCode,
+        ]);
+    });
+
+    api.modifyHTMLTags(
+      async (
+        { headTags, bodyTags },
+        { environment, filename, assetPrefix },
+      ) => {
+        const entries = Object.keys(environment.entry);
+        const { config } = environment;
+
+        const isExclude = options.excludeEntries.find((item: string) => {
+          if (!entries.includes(item)) {
+            logger.error(`convertToRem: can't find the entryName: ${item}`);
+            return false;
+          }
+
+          const reg = new RegExp(`(/${item}/index.html)|(/${item}.html)`, 'gi');
+          return reg.test(filename);
+        });
+
+        if (isExclude) {
+          return { headTags, bodyTags };
         }
 
-        const { AutoSetRootFontSizePlugin } = await import(
-          './AutoSetRootFontSizePlugin'
-        );
+        const scriptTag = {
+          tag: 'script',
+          attrs: {},
+        };
 
-        const entries = Object.keys(chain.entryPoints.entries() || {});
-        const { config } = environment;
-        const distDir = config.output.distPath.js;
+        let injectPosition = 0;
+        // should inject rem runtime code after meta tags
+        headTags.forEach((tag, index) => {
+          if (tag.tag === 'meta') {
+            injectPosition = index + 1;
+          }
+        });
 
-        chain
-          .plugin(CHAIN_ID.PLUGIN.AUTO_SET_ROOT_SIZE)
-          .use(AutoSetRootFontSizePlugin, [
-            userOptions,
-            entries,
-            HtmlPlugin,
-            distDir,
-            config.html.scriptLoading,
-          ]);
+        if (options.inlineRuntime) {
+          headTags.splice(injectPosition, 0, {
+            ...scriptTag,
+            children: await getRuntimeCode(),
+          });
+        } else {
+          const url = ensureAssetPrefix(
+            getScriptPath(config.output.distPath.js),
+            assetPrefix,
+          );
+
+          const attrs: Record<string, string> = {
+            ...scriptTag.attrs,
+            src: url,
+          };
+
+          if (config.html.scriptLoading === 'defer') {
+            attrs.defer = '';
+          }
+          if (config.html.scriptLoading === 'module') {
+            attrs.type = 'module';
+          }
+
+          headTags.splice(injectPosition, 0, {
+            ...scriptTag,
+            attrs,
+          });
+        }
+
+        return {
+          headTags,
+          bodyTags,
+        };
       },
     );
   },
